# feat: add indexes to improve resource filtering and batch selection