## Add indexes to improve resource filtering and batch selection

### Description
Added two new indexes to the catalog table to improve query performance:

1. `status_deleted_idx` on `(status, deleted)`:

- Improves performance of resource filtering queries from ~4.5s to ~0.5s
- Used in `Resource.get_excluded_clause()` which filters resources for crawling

2. `last_check_priority_idx` on `(last_check, priority)`:

- Improves performance of batch selection queries
- Used in `select_batch_resources_to_check()` for resource processing

### Testing
Performance was tested using `EXPLAIN ANALYZE` on the main queries:
Resource filtering query (status/deleted): 4.5s → 0.5s
Batch selection query: consistently performs at ~0.1s with caching

### Impact
These indexes will improve the overall performance of the crawling pipeline without any code changes required. The trade-off of additional disk space for indexes is justified by the significant performance gains in these frequently-used queries.